src/x_to_nwb/conversion.py: Add generic channel selection options for ABF
Up to know the list of written AD channels into NWB was a mixture of
hardcoded via ABFConverter.adcNamesWithRealData and the possibility to
extend that list either to all using --outputFeedbackChannel or
partially with --realDataChannel.

This approach does not work for arbitrary ABF files.

We now have new options --includeChannel and --discardChannel which
allows total control over the outputted AD channels.